Refactroed Game to remove scanner within and moved to GameController to handle I/O
